+++ tests/Fixtures/Canvas/Models/User.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Canvas\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
+class User extends Model
+{
+    use SoftDeletes;
+
+    protected $table = 'canvas_users';
+
+    protected $keyType = 'string';
+
+    public $incrementing = false;
+
+    protected $guarded = [];
+}